# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Multi-Warehouse & Multi-Country Inventory Management System API",
        "description": "This documentation describes all REST endpoints exposed by the Inventory Management backend. All requests (except authentication) are protected with JWT (Bearer) tokens.",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Primary API server"
        }
    ],
    "paths": {
...
        "/reports/low-stock": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Daily low stock report",
                "operationId": "260014f05132d4e794a4e2ea7720a3fb",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LowStockReportResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List products",
                "operationId": "080f057905dc32d2f11065d606dc4cd6",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create product",
                "operationId": "eddf8c7847239ddc1040daadd814e21b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products/{product}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product",
                "operationId": "c5ec5b4fb4c3790c73e9849244665f45",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update product",
                "operationId": "f7375aee24e5fccaf1567852e3492363",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete product",
                "operationId": "fb827fecd27dd65fa2b514023be16841",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/suppliers": {
            "get": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "List suppliers",
                "operationId": "5d29c6f6c1ac2bd100e650844ae63ab2",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SupplierResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Create supplier",
                "operationId": "51d12a3c668c583e720a6cee45e0c31e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SupplierRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SupplierResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/suppliers/{supplier}": {
            "get": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Get supplier",
                "operationId": "f28f945e0e67308feb10fd1c67899074",
                "parameters": [
                    {
                        "name": "supplier",
                        "in": "path",
                        "description": "Supplier ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SupplierResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Update supplier",
                "operationId": "9b774aa678dcb1a462cd6c8b301e6d3a",
                "parameters": [
                    {
                        "name": "supplier",
                        "in": "path",
                        "description": "Supplier ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SupplierRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SupplierResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Delete supplier",
                "operationId": "958a6acc282e0be5e45e073837bd9bf0",
                "parameters": [
                    {
                        "name": "supplier",
                        "in": "path",
                        "description": "Supplier ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductRequest": {
                "title": "Product Request",
                "description": "Product request body data",
                "required": [
                    "name",
                    "sku",
                    "price"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "iPhone 14"
                    },
                    "sku": {
                        "type": "string",
                        "maxLength": 60,
                        "example": "IPH-14-128BLK"
                    },
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Apple iPhone 14 128GB Black"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 999.99
                    }
                },
                "type": "object"
            },
            "SupplierRequest": {
                "title": "Supplier Request",
                "description": "Supplier request body data",
                "required": [
                    "name",
                    "contact_info",
                    "address"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Acme Supplies"
                    },
                    "contact_info": {
                        "type": "object",
                        "example": {
                            "phone": "+1-555-1234",
                            "email": "sales@acme.com"
                        }
                    },
                    "address": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "123 Industrial Rd, NY"
                    }
                },
                "type": "object"
            },
...
            "LowStockReportResource": {
                "title": "Low Stock Report Resource",
                "description": "Low stock report representation",
                "properties": {
                    "productName": {
                        "type": "string",
                        "example": "iPhone 14"
                    },
                    "SKU": {
                        "type": "string",
                        "example": "IPH-14-128BLK"
                    },
                    "currentQuantity": {
                        "type": "integer",
                        "example": 10
                    },
                    "minQuantity": {
                        "type": "integer",
                        "example": 20
                    },
                    "warehouseLocation": {
                        "type": "string",
                        "example": "New York"
                    },
                    "country": {
                        "type": "string",
                        "example": "United States"
                    },
                    "supplierContactInfo": {
                        "type": "object",
                        "example": {
                            "phone": "+1-555-1234",
                            "email": "sales@acme.com"
                        }
                    }
                },
                "type": "object"
            },
            "ProductResource": {
                "title": "Product Resource",
                "description": "Product resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "iPhone 14"
                    },
                    "sku": {
                        "type": "string",
                        "example": "IPH-14-128BLK"
                    },
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Apple iPhone 14 128GB Black"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 999.99
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "SupplierResource": {
                "title": "Supplier Resource",
                "description": "Supplier resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Acme Supplies"
                    },
                    "contact_info": {
                        "type": "object",
                        "example": {
                            "phone": "+1-555-1234",
                            "email": "sales@acme.com"
                        }
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Industrial Rd, NY"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}